// src/app/models/book.ts
export class Book {
    title: String = '1984';
    description: String = 'Big Brother';
    thumbnail: String = 'https://images.unsplash.com/photo-1622609184693-58079bb6742f?crop=entropy&cs=tinysrgb&fit=max&fm=jpg&ixid=MnwxfDB8MXxyYW5kb218MHx8fHx8fHx8MTYzNjY1OTc1MQ&ixlib=rb-1.2.1&q=80&utm_campaign=api-credit&utm_medium=referral&utm_source=unsplash_source&w=1080';
    author: String = 'George Orwell';
    type: String = 'Fiction';
}